feat: Add partners section to home page
This commit adds a new partners section to the home page. The `PartnersSection` component is created in the `partners/page.tsx` file, which displays a list of partners with their logos. The partners data is fetched from the API using the `getPartnersData` function in the `api.ts` file. The `PartnersSection` component is then added to the main content of the home page in the `Home` component.